requirements.txt
# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Crawler
scrapy>=2.11.0

# JS rendering (optional)
playwright>=1.40.0

# Text extraction
trafilatura>=1.6.0
readability-lxml>=0.8.1
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Utilities
python-dateutil>=2.8.0
